requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
pandas
numpy
scikit-learn==1.5.0
matplotlib==3.9.0
plotly==5.22.0
python-dotenv==1.0.1
netifaces==0.11.0
psutil==5.9.8